fix: remove eslint-disable comments and defensive try-catch
- Move service imports to db-test-seeders and re-export through
  api-test-helpers to eliminate 3 eslint-disable comments
- Remove defensive try-catch in credit consumer (let errors propagate
  to dispatch layer's Promise.allSettled)
- Remove unused _userId parameter from getChatThreadMessages
